Add create controller for comments

+++ src/main/java/com/victorku/musiccloud/web/model/CommentsDTO.java
@@ -5,12 +5,12 @@ public class CommentsDTO {
 
     private Long id;
     private String text;
-    private Long orderComments;
+    private Integer orderComments;
 
     public CommentsDTO() {
     }
 
-    public CommentsDTO(Long id, String text, Long orderComments) {
+    public CommentsDTO(Long id, String text, Integer orderComments) {
         this.id = id;
         this.text = text;
         this.orderComments = orderComments;
@@ -32,11 +32,11 @@ public class CommentsDTO {
         this.text = text;
     }
 
-    public Long getOrderComments() {
+    public Integer getOrderComments() {
         return orderComments;
     }
 
-    public void setOrderComments(Long orderComments) {
+    public void setOrderComments(Integer orderComments) {
         this.orderComments = orderComments;
     }
 }
